requirements.txt
instagrapi==2.1.2
schedule==1.2.0
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0